requirements.txt
# LINUX REQUIREMENTS
# sudo apt-get install libpq-dev python3-dev
# sudo apt-get install postgresql postgresql-contrib

# CREATING DB AND USER FOR POSTGRESQL
# sudo -u postgres psql
# CREATE DATABASE gamesdb;
# CREATE USER gamesdbuser WITH ENCRYPTED PASSWORD 'gamesdbuser';
# ALTER ROLE gamesdbuser SET client_encoding TO 'utf8';
# ALTER ROLE gamesdbuser SET default_transaction_isolation TO 'read committed';
# ALTER ROLE gamesdbuser SET timezone TO 'UTC';
# GRANT ALL PRIVILEGES ON DATABASE gamesdb TO gamesdbuser;
# \q

# PIP REQUIREMENTS
django~=3.2
djangorestframework~=3.12.4
django-environ

requests~=2.25.1
beautifulsoup4~=4.9.3
lxml

psycopg2
psycopg2-binary
